feat: add home drawer image previews and search functionality
- Implemented image preview caching and loading for conversations in the home drawer.
- Added models for drawer shortcuts and conversation search results.
- Developed search functionality to filter conversations based on user input.
- Enhanced conversation loading with background image preview preloading.
- Introduced methods for managing conversation snapshots and image preview states.
