+++ ui/lib/features/home/pages/command_overlay/widgets/cards/agent_tool_summary_card.dart
@@ -52,6 +52,13 @@ class AgentToolSummaryCard extends StatelessWidget {
     final titleColor = context.isDarkTheme
         ? palette.textPrimary
         : visualProfile.primaryTextColor;
+    final titleStyle = TextStyle(
+      color: titleColor,
+      fontSize: 12,
+      fontWeight: FontWeight.w500,
+      letterSpacing: 0,
+      height: 1.15,
+    );
 
     final tooltipLines = <String>[title];
     if (preview.isNotEmpty && preview != title) {
@@ -95,17 +102,17 @@ class AgentToolSummaryCard extends StatelessWidget {
                         ),
                         const SizedBox(width: 8),
                         Flexible(
-                          child: Text(
-                            title,
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                            style: TextStyle(
-                              color: titleColor,
-                              fontSize: 12,
-                              fontWeight: FontWeight.w500,
-                              height: 1.15,
-                            ),
-                          ),
+                          child: status == 'running'
+                              ? _FlowingToolTitleText(
+                                  text: title,
+                                  style: titleStyle,
+                                )
+                              : Text(
+                                  title,
+                                  maxLines: 1,
+                                  overflow: TextOverflow.ellipsis,
+                                  style: titleStyle,
+                                ),
                         ),
                         const SizedBox(width: 8),
                         Container(
@@ -136,6 +143,81 @@ class AgentToolSummaryCard extends StatelessWidget {
           ),
         ),
       ),
+    );
+  }
+}
+
+class _FlowingToolTitleText extends StatefulWidget {
+  const _FlowingToolTitleText({required this.text, required this.style});
+
+  final String text;
+  final TextStyle style;
+
+  @override
+  State<_FlowingToolTitleText> createState() => _FlowingToolTitleTextState();
+}
+
+class _FlowingToolTitleTextState extends State<_FlowingToolTitleText>
+    with SingleTickerProviderStateMixin {
+  late final AnimationController _controller = AnimationController(
+    vsync: this,
+    duration: const Duration(milliseconds: 1800),
+  )..repeat();
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final child = Text(
+      widget.text,
+      maxLines: 1,
+      overflow: TextOverflow.ellipsis,
+      style: widget.style,
+    );
+    if (MediaQuery.disableAnimationsOf(context)) {
+      return child;
+    }
+
+    final baseColor = widget.style.color ?? context.omniPalette.textPrimary;
+    final highlightColor = context.isDarkTheme
+        ? Colors.white.withValues(alpha: 0.96)
+        : Colors.white.withValues(alpha: 0.92);
+
+    return AnimatedBuilder(
+      animation: _controller,
+      child: child,
+      builder: (context, child) {
+        return ShaderMask(
+          blendMode: BlendMode.srcIn,
+          shaderCallback: (bounds) {
+            final textWidth = bounds.width <= 0 ? 1.0 : bounds.width;
+            final shimmerWidth = (textWidth * 0.72)
+                .clamp(52.0, 180.0)
+                .toDouble();
+            final travelDistance = textWidth + shimmerWidth;
+            final shimmerLeft =
+                bounds.left - shimmerWidth + travelDistance * _controller.value;
+            return LinearGradient(
+              begin: Alignment.centerLeft,
+              end: Alignment.centerRight,
+              colors: [baseColor, highlightColor, baseColor],
+              stops: const [0.08, 0.5, 0.92],
+            ).createShader(
+              Rect.fromLTWH(
+                shimmerLeft,
+                bounds.top,
+                shimmerWidth,
+                bounds.height,
+              ),
+            );
+          },
+          child: child,
+        );
+      },
     );
   }
 }
