+++ ui/lib/features/home/widgets/home_drawer_actions.dart
@@ -1,0 +1,376 @@
+// ignore_for_file: invalid_use_of_protected_member
+
+part of 'home_drawer.dart';
+
+extension _HomeDrawerActions on HomeDrawerState {
+  bool get _shouldCloseOnNavigate => widget.closeOnNavigate && !widget.embedded;
+
+  void _maybeCloseDrawer() {
+    if (!_shouldCloseOnNavigate || !Navigator.of(context).canPop()) {
+      return;
+    }
+    Navigator.pop(context);
+  }
+
+  void _openThreadTarget(ConversationThreadTarget target) {
+    if (widget.embedded && widget.onThreadTargetSelected != null) {
+      widget.onThreadTargetSelected!(target);
+      return;
+    }
+    _maybeCloseDrawer();
+    GoRouterManager.push(
+      '/home/chat',
+      extra: target,
+      queryParams: _threadTargetQueryParams(target),
+    );
+  }
+
+  Map<String, dynamic> _threadTargetQueryParams(
+    ConversationThreadTarget target,
+  ) {
+    return <String, dynamic>{
+      'conversationId': target.conversationId?.toString() ?? 'new',
+      'mode': target.mode.storageValue,
+      'requestKey':
+          target.requestKey ?? DateTime.now().microsecondsSinceEpoch.toString(),
+    };
+  }
+
+  void _navigateTo(String route) {
+    _maybeCloseDrawer();
+    GoRouterManager.push(route);
+  }
+
+  void _openNewConversation() {
+    _openThreadTarget(
+      ConversationThreadTarget.newConversation(
+        mode: widget.newConversationMode,
+        requestKey: DateTime.now().microsecondsSinceEpoch.toString(),
+      ),
+    );
+  }
+
+  Future<void> _triggerDeleteHaptic() async {
+    try {
+      final enabled = await CacheUtil.getBool(
+        'app_vibrate',
+        defaultValue: true,
+      );
+      if (!enabled) {
+        return;
+      }
+      await HapticFeedback.mediumImpact();
+    } catch (error) {
+      debugPrint('[HomeDrawer] failed to trigger delete haptic: $error');
+    }
+  }
+
+  void _replaceConversationInState(ConversationModel updatedConversation) {
+    final allConversations = List<ConversationModel>.from(_allConversations);
+    final allIndex = allConversations.indexWhere(
+      (item) => item.threadKey == updatedConversation.threadKey,
+    );
+    if (allIndex >= 0) {
+      allConversations[allIndex] = updatedConversation;
+    }
+
+    final searchResults = List<_ConversationSearchResult>.from(_searchResults);
+    final searchIndex = searchResults.indexWhere(
+      (item) => item.conversation.threadKey == updatedConversation.threadKey,
+    );
+    if (searchIndex >= 0) {
+      searchResults[searchIndex] = searchResults[searchIndex].copyWith(
+        conversation: updatedConversation,
+      );
+    }
+
+    _conversationSearchCache.remove(updatedConversation.threadKey);
+    _allConversations = allConversations;
+    _searchResults = searchResults;
+    _syncConversationSnapshotCache();
+  }
+
+  void _removeConversationFromState(ConversationModel conversation) {
+    _allConversations = List<ConversationModel>.from(_allConversations)
+      ..removeWhere((item) => item.threadKey == conversation.threadKey);
+    _searchResults = List<_ConversationSearchResult>.from(_searchResults)
+      ..removeWhere(
+        (item) => item.conversation.threadKey == conversation.threadKey,
+      );
+    _conversationSearchCache.remove(conversation.threadKey);
+    _syncConversationSnapshotCache();
+  }
+
+  void _startEditingTitle(ConversationModel conversation) {
+    if (_busyConversationKeys.contains(conversation.threadKey)) {
+      return;
+    }
+    final title = _resolveConversationTitle(conversation);
+    _titleEditingController.text = title;
+    _titleEditingController.selection = TextSelection(
+      baseOffset: 0,
+      extentOffset: title.length,
+    );
+    setState(() {
+      _editingThreadKey = conversation.threadKey;
+    });
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _titleEditingFocusNode.requestFocus();
+    });
+  }
+
+  void _handleTitleEditingFocusChanged() {
+    if (!_titleEditingFocusNode.hasFocus && _editingThreadKey != null) {
+      _commitTitleEdit();
+    }
+  }
+
+  Future<void> _commitTitleEdit() async {
+    final threadKey = _editingThreadKey;
+    if (threadKey == null) return;
+
+    final newTitle = _titleEditingController.text.trim();
+    final conversation = _allConversations
+        .cast<ConversationModel?>()
+        .firstWhere((c) => c!.threadKey == threadKey, orElse: () => null);
+
+    setState(() {
+      _editingThreadKey = null;
+    });
+
+    if (conversation == null) return;
+
+    final oldTitle = _resolveConversationTitle(conversation);
+    if (newTitle.isEmpty || newTitle == oldTitle) return;
+
+    final updated = conversation.copyWith(title: newTitle);
+    setState(() {
+      _replaceConversationInState(updated);
+    });
+    _persistCurrentConversationSnapshot();
+
+    final success = await ConversationService.updateConversationTitle(
+      conversationId: conversation.id,
+      newTitle: newTitle,
+      mode: conversation.mode,
+    );
+
+    if (!mounted) return;
+    if (!success) {
+      setState(() {
+        _replaceConversationInState(conversation);
+      });
+      _persistCurrentConversationSnapshot();
+      showToast(context.trLegacy('重命名失败'), type: ToastType.error);
+    }
+  }
+
+  void _openConversationFromDrawer(ConversationModel conversation) {
+    if (_busyConversationKeys.contains(conversation.threadKey)) {
+      return;
+    }
+    _openThreadTarget(
+      ConversationThreadTarget.existing(
+        conversationId: conversation.id,
+        mode: conversation.mode,
+      ),
+    );
+  }
+
+  Future<void> _deleteConversation(ConversationModel conversation) async {
+    if (_busyConversationKeys.contains(conversation.threadKey)) {
+      return;
+    }
+
+    final originalIndex = _allConversations.indexWhere(
+      (item) => item.id == conversation.id,
+    );
+    if (originalIndex < 0) {
+      return;
+    }
+
+    setState(() {
+      _busyConversationKeys.add(conversation.threadKey);
+      _removeConversationFromState(conversation);
+    });
+    _persistCurrentConversationSnapshot();
+
+    final deleted = await ConversationService.deleteConversation(
+      conversation.id,
+      mode: conversation.mode,
+    );
+    if (!mounted) {
+      return;
+    }
+    if (deleted) {
+      unawaited(_triggerDeleteHaptic());
+    }
+
+    setState(() {
+      _busyConversationKeys.remove(conversation.threadKey);
+      if (!deleted) {
+        final restoredIndex = originalIndex <= _allConversations.length
+            ? originalIndex
+            : _allConversations.length;
+        _allConversations = List<ConversationModel>.from(_allConversations)
+          ..insert(restoredIndex, conversation);
+        _syncConversationSnapshotCache();
+        if (_isSearchActive) {
+          _scheduleConversationSearch(immediate: true);
+        }
+      }
+    });
+    _persistCurrentConversationSnapshot();
+
+    showToast(
+      deleted ? context.trLegacy('已删除') : context.trLegacy('删除失败'),
+      type: deleted ? ToastType.success : ToastType.error,
+    );
+  }
+
+  Future<void> _archiveConversation(ConversationModel conversation) async {
+    if (_busyConversationKeys.contains(conversation.threadKey)) {
+      return;
+    }
+
+    final originalIndex = _allConversations.indexWhere(
+      (item) => item.threadKey == conversation.threadKey,
+    );
+    if (originalIndex < 0) {
+      return;
+    }
+
+    final originalConversation = _allConversations[originalIndex];
+    final archivedConversation = originalConversation.copyWith(
+      isArchived: true,
+    );
+
+    setState(() {
+      _busyConversationKeys.add(conversation.threadKey);
+      _replaceConversationInState(archivedConversation);
+    });
+    _persistCurrentConversationSnapshot();
+
+    final archived = await ConversationService.archiveConversation(
+      originalConversation,
+    );
+    if (!mounted) {
+      return;
+    }
+
+    setState(() {
+      _busyConversationKeys.remove(conversation.threadKey);
+      if (!archived) {
+        _replaceConversationInState(originalConversation);
+      }
+    });
+    _persistCurrentConversationSnapshot();
+
+    showToast(
+      archived ? context.trLegacy('已归档') : context.trLegacy('归档失败'),
+      type: archived ? ToastType.success : ToastType.error,
+    );
+  }
+
+  Future<void> _unarchiveConversation(ConversationModel conversation) async {
+    if (_busyConversationKeys.contains(conversation.threadKey)) {
+      return;
+    }
+
+    final originalIndex = _allConversations.indexWhere(
+      (item) => item.threadKey == conversation.threadKey,
+    );
+    if (originalIndex < 0) {
+      return;
+    }
+
+    final originalConversation = _allConversations[originalIndex];
+    final restoredConversation = originalConversation.copyWith(
+      isArchived: false,
+    );
+
+    setState(() {
+      _busyConversationKeys.add(conversation.threadKey);
+      _replaceConversationInState(restoredConversation);
+    });
+    _persistCurrentConversationSnapshot();
+
+    final restored = await ConversationService.unarchiveConversation(
+      originalConversation,
+    );
+    if (!mounted) {
+      return;
+    }
+
+    setState(() {
+      _busyConversationKeys.remove(conversation.threadKey);
+      if (!restored) {
+        _replaceConversationInState(originalConversation);
+      }
+    });
+    _persistCurrentConversationSnapshot();
+
+    showToast(
+      restored ? context.trLegacy('已取消归档') : context.trLegacy('取消归档失败'),
+      type: restored ? ToastType.success : ToastType.error,
+    );
+  }
+
+  List<ConversationSlideAction> _buildDrawerActions(
+    ConversationModel conversation,
+  ) {
+    return [
+      ConversationSlideAction(
+        onPressed: () => _deleteConversation(conversation),
+        backgroundColor: AppColors.alertRed,
+        child: Center(
+          child: SvgPicture.asset(
+            'assets/memory/memory_delete.svg',
+            width: HomeDrawerState._conversationActionIconSize,
+            height: HomeDrawerState._conversationActionIconSize,
+            colorFilter: const ColorFilter.mode(Colors.white, BlendMode.srcIn),
+          ),
+        ),
+      ),
+      ConversationSlideAction(
+        onPressed: () => conversation.isArchived
+            ? _unarchiveConversation(conversation)
+            : _archiveConversation(conversation),
+        backgroundColor: context.isDarkTheme
+            ? Color.lerp(
+                context.omniPalette.surfaceElevated,
+                context.omniPalette.accentPrimary,
+                0.3,
+              )!
+            : AppColors.buttonPrimary,
+        borderRadius: HomeDrawerState._drawerTrailingActionRadius,
+        child: Center(
+          child: conversation.isArchived
+              ? Icon(
+                  Icons.unarchive_outlined,
+                  size: HomeDrawerState._conversationActionIconSize,
+                  color: Colors.white,
+                )
+              : SvgPicture.asset(
+                  'assets/home/archive_icon.svg',
+                  width: HomeDrawerState._conversationActionIconSize,
+                  height: HomeDrawerState._conversationActionIconSize,
+                  colorFilter: const ColorFilter.mode(
+                    Colors.white,
+                    BlendMode.srcIn,
+                  ),
+                ),
+        ),
+      ),
+    ];
+  }
+
+  String _resolveConversationTitle(ConversationModel conversation) {
+    final title = conversation.title.trim();
+    if (title.isNotEmpty) {
+      return title;
+    }
+    final summary = (conversation.summary ?? '').trim();
+    return summary.isNotEmpty ? summary : context.trLegacy('未命名对话');
+  }
+}
